Refactored SizeExtensions to avoid overflow
