+++ src/Nethermind/Nethermind.Core/Extensions/IntExtensions.cs
@@ -20,69 +20,6 @@ using Nethermind.Int256;
 
 namespace Nethermind.Core.Extensions
 {
-    public static class SizeExtensions
-    {
-        public static long GB(this int @this)
-        {
-            return @this * 1_000_000_000L;
-        }
-
-        public static long MB(this int @this)
-        {
-            return @this * 1_000_000L;
-        }
-
-        public static long KB(this int @this)
-        {
-            return @this * 1_000L;
-        }
-
-        public static long GiB(this int @this)
-        {
-            return @this * 1024L * 1024L * 1024L;
-        }
-
-        public static long MiB(this int @this)
-        {
-            return @this * 1024L * 1024L;
-        }
-
-        public static long KiB(this int @this)
-        {
-            return @this * 1024L;
-        }
-
-        public static long GB(this long @this)
-        {
-            return ((int)@this).GB();
-        }
-
-        public static long MB(this long @this)
-        {
-            return ((int)@this).MB();
-        }
-
-        public static long KB(this long @this)
-        {
-            return ((int)@this).KB();
-        }
-
-        public static long GiB(this long @this)
-        {
-            return ((int)@this).GiB();
-        }
-
-        public static long MiB(this long @this)
-        {
-            return ((int)@this).MiB();
-        }
-
-        public static long KiB(this long @this)
-        {
-            return ((int)@this).KiB();
-        }
-    }
-
     public static class IntExtensions
     {
         public static string ToHexString(this int @this)
